requirements.txt
asgiref==3.3.1
astroid==2.9.3
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.7
django-cors-headers==3.7.0
django-rest-framework==0.1.0
django-rest-passwordreset==1.2.1
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.1
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
openapi-codec==1.3.2
packaging==21.0
platformdirs==2.4.1
PyJWT==1.7.1
pylint==2.12.2
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
simplejson==3.17.5
sqlparse==0.4.1
toml==0.10.2
typing_extensions==4.0.1
uritemplate==3.0.1
urllib3==1.26.6
wrapt==1.13.3
